transformer: kernel_source_packager: New RepoWorktree class
To be able to re-use builder vm and share it within a team we need to use
git-worktree to separate working directories. The new class introduces a new
`location.type`: worktree which if given `location.worktree_name` would add a
new worktree at `location.path`, and set up a repo and branch as specified by
either `location.worktree_repo` and `location.worktree_ref` or by
`location.repo` and `location.ref`.
If a worktree_name is not set, then the class would behave similar to
`location.type`: repo.

Signed-off-by: Piyush Sachdeva <psachdeva@microsoft.com>